+++ app/signup/EmailAuth.tsx
@@ -105,13 +105,6 @@ function EmailAuth() {
     disabled: !(formValue.code.length > 0),
   };
 
-  const ImageProps = {
-    src: "/images/svg/icon_X.svg",
-    alt: "NEXT ROOM",
-    width: 28,
-    height: 28,
-  };
-
   const errorMessage = isError && error?.response?.data?.message;
 
   useEffect(() => {
@@ -141,7 +134,6 @@ function EmailAuth() {
     timeLeft,
     minutes,
     second,
-    ImageProps,
     formProps,
     adminCodeProps,
     buttonProps,
