+++ app/signup/EmailAuthView.tsx
@@ -1,10 +1,8 @@
 import React, { useEffect } from "react";
-import Image from "next/image";
 
 import Loader from "@/components/Loader/Loader";
 import SnackBar from "@/components/SnackBar/SnackBar";
 import { useSnackBarInfo } from "@/components/atoms/snackBar.atom";
-import { useRouter } from "next/navigation";
 import { getAnalytics, logEvent } from "firebase/analytics";
 import CodeInput from "./CodeInput";
 import "@/apis/firebase";
@@ -18,7 +16,6 @@ function EmailAuthView(props: Props) {
     minutes,
     second,
     timeLeft,
-    ImageProps,
     formProps,
     signUpState,
     buttonProps,
@@ -29,7 +26,6 @@ function EmailAuthView(props: Props) {
 
   const [snackInfo, setSnackBarInfo] = useSnackBarInfo();
   const analytics = getAnalytics();
-  const router = useRouter();
 
   logEvent(analytics, "screen_view", {
     firebase_screen: "sign_up_2",
@@ -44,49 +40,36 @@ function EmailAuthView(props: Props) {
   }, [setSnackBarInfo, snackInfo]);
 
   return (
-    <>
-      <S.Wrapper>
-        {isLoading && <Loader />}
-        <S.Header
-          onClick={() => {
-            router.back();
-            window.close();
-          }}
-        >
-          <Image {...ImageProps} />
-        </S.Header>
-      </S.Wrapper>
+    <S.Cont>
+      {isLoading && <Loader />}
+      <S.Title>
+        {signUpState.email}으로
+        <br />
+        전송된 인증번호를 입력해 주세요.
+      </S.Title>
+      <S.SubTitle>
+        남은 시간 {minutes}:{second}
+      </S.SubTitle>
 
-      <S.Cont>
-        <S.Title>
-          {signUpState.email}으로
-          <br />
-          전송된 인증번호를 입력해 주세요.
-        </S.Title>
-        <S.SubTitle>
-          남은 시간 {minutes}:{second}
-        </S.SubTitle>
-
-        <S.StyledBox {...formProps}>
-          {/* <TextField {...adminCodeProps} /> */}
-          <CodeInput {...formProps} {...buttonProps} disabled={timeLeft} />
-          <S.ReRequest>
-            <p>이메일이 오지 않았다면? </p>
-            <button type="button" {...ReRequestButtonProps}>
-              재인증 요청
-            </button>
-          </S.ReRequest>
-          <S.ServerErrorMessage>{errorMessage}</S.ServerErrorMessage>
-        </S.StyledBox>
-        <SnackBar
-          open={snackInfo.isOpen}
-          ment={snackInfo.message}
-          vertical="top"
-          horizontal="center"
-          handleClose={() => setSnackBarInfo({ ...snackInfo, isOpen: false })}
-        />
-      </S.Cont>
-    </>
+      <S.StyledBox {...formProps}>
+        {/* <TextField {...adminCodeProps} /> */}
+        <CodeInput {...formProps} {...buttonProps} disabled={timeLeft} />
+        <S.ReRequest>
+          <p>이메일이 오지 않았다면? </p>
+          <button type="button" {...ReRequestButtonProps}>
+            재인증 요청
+          </button>
+        </S.ReRequest>
+        <S.ServerErrorMessage>{errorMessage}</S.ServerErrorMessage>
+      </S.StyledBox>
+      <SnackBar
+        open={snackInfo.isOpen}
+        ment={snackInfo.message}
+        vertical="top"
+        horizontal="center"
+        handleClose={() => setSnackBarInfo({ ...snackInfo, isOpen: false })}
+      />
+    </S.Cont>
   );
 }
